requirements.txt
requests
python-dotenv
flask-mail
APScheduler
schedule
celery 
redis
waitress
gunicorn
Flask-Dance
google-auth
flask-cors
SQLAlchemy
Flask-SQLAlchemy
Flask-Login
bcrypt
flask-jwt-extended
psycopg2-binary
Flask-APScheduler